refactoring models
