+++ packages/tracing-processing/test/costants.ts
@@ -1,9 +1,7 @@
 import { generateId, SavePurposeErrorDto } from "pagopa-interop-tracing-models";
-import {
-  TracingRecordSchema,
-  Eservice,
-  TracingFromCsv,
-} from "../src/models/messages.js";
+import { Eservice } from "../src/models/csv.js";
+import { TracingRecordSchema } from "../src/models/db.js";
+import { TracingFromS3Path } from "../src/models/tracing.js";
 
 export const mockTracingRecords: TracingRecordSchema[] = [
   {
@@ -229,7 +227,7 @@ const tenant_id = "123e4567-e89b-12d3-a456-426614174001";
 const purposeId = generateId() as string;
 const purposeId2 = generateId() as string;
 
-export const mockMessage: TracingFromCsv = {
+export const mockMessage: TracingFromS3Path = {
   tenantId: tenant_id,
   date: "2024-12-12",
   version: 1,
